+++ magda-admin-api/src/connectorObjName.ts
@@ -1,0 +1,2 @@
+const connectorObjName = (id: string) => `connector-${id}`;
+export default connectorObjName;