+++ CHANGES.md
@@ -53,6 +53,7 @@
 - #3317 Use multi-stage builds to build magda-scss-compile
 - #3337 Release docker images & helm charts to Github container registry as well
 - #3349 secure content API with new auth model
+- #3350 secure admin API (managing connectors) with new auth model
 
 ## 1.3.1